Add schema evolution for Parquet Hive table
Known limitation (Not an issue with Criteo usage of the metastore):
If metastore partition schema doesn't advertising a column that exist in the parquet file.
And the table schema advertise it.
Then the column valeus are read and returned.
Thus metastore can't be used to hide column in a subpart of a table partitions.